Refactor app to use flask factory and blueprints
Why these changes are being introduced:
The app was previously written using a simplified app creation method
that caused multiple issues: one around imports, and another around
using moto3 for S3 testing (related to the imports and initialization
of the app). This has been preventing dependency updates for some time.

How this addresses that need:
* Refactors the app to use Flask factories and blueprints, which is now
  the standard Flask app model
* Updates S3 connection initialization to happen within an app request
  function, providing access to the app config for S3 secrets info
* Updates the configuration to use env-based classes to simplify
  configuration across multiple environments

Side effects of this change:
We will be able to update all dependencies (in a future PR).