Update to use Alma API
Why these changes are being introduced:
The application currently uses the Aleph API to retrieve metadata for
ebook items. This updates the application to use the Alma API instead.

How this addresses that need:
* Updates configuration and query code to use Alma API key and URL
* Updates test fixtures and corresponding tests to use Alma sample
  records
* Updates documentation in README

Side effects of this change:
None, but the full application will not work until the file names for
the ebook files (stored in S3) are updated to use the Alma MMSIDs
instead of the Aleph bib record numbers. There is already a ticket for
this.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IMP-2010