Temporary fix to broken SAML auth
Why these changes are being introduced:
For reasons that still aren't entirely clear, SAML auth isn't working in
prod only. This update uses a forked version of the python3-saml lib
with a minor edit that fixes the issue until we can fully understand and
resolve it.

How this addresses that need:
* Updates python3-saml package to a locally-maintained forked.
* Updates SAML settings to debug in all environments.
* Updates AWS signing to v4, which is required now that we use KMS key
  encryption for our S3 buckets.
* Updates dependencies.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-361