Update Ebooks app with new MIT branding
Why these changes are being introduced:
* The Institute recently updated its logo and branding guidelines, which
  we need to keep current.

How this addresses that need:
* Use new logos and favicon files from CDN and remove image files from repo
* Add new 'Contact us' link
* Add tool name to header
* Apply consistent padding and width to logos
* Remove “Search Hours & locations Borrow & request Research support About us” text/links
* Add Docker compose yaml file for local IdP

Side effects of this change:
* Minimal updates were made to the HTML and CSS files, enough to
render the expected header and footer in accordance with the updated
branding guidelines. The code and syntax in these files can definitely
be improved, but it is out of scope for this ticket.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/PW-77