Update Talisman configurations to load external resources
Why these changes are being introduced:
* While reviewing the MIT branding updates in production, a number
of errors appeared in the console. The default Talisman content
security policy configs are extremely strict and prevents loading
any resources that are not in the same domain as the application.
This resulted in the MIT images not rendering for the application
(the images are loaded from our CDN). It also prevented running
the script for Google Analytics.

How this addresses that need:
* Configure trusted domains for Talisman
* Use 'nonce' to enable execution of in-line script for Google Analytics

Side effects of this change:
* Updated Content Security Policy for app

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/PW-77